Bug fix h5.get_attributes takes iteration string
This is also a reminder that I shall not have so many repeated attributes
of the iteration setup on each iteration. Iterations shall be all the
same only the seeding/output green function should be the
different. Have a unique dictionary for the setup not per iteration, an
attribute to the group.