Real Frequency dos dimer temperature evolution
This update comes because in 51365ec834ba05e1360d8a0053b62610b99aadf5 I
changed the infinitesimal positive constant that used to give the causal
Green Function for the Weiss field. This changed the end results to what
I had registered for last week and I could not understand them. First I
thought I originated from the update in numpy and scipy. But was not the
case.

In the end it returns to the fact that this positive constant of
causality has to at least twice delta omega. This is because in the
convolutions I get a shift of dw, and I have 2 convolutions. So to make
sure I get the zero Dirac delta divergence captured by a pole I need i2dw as
the constant of causality. More than that changes the results. I don't
really have a good argument on which is a better result, but I shall try
to keep this constant as small as possible, just because in theory it is
like that, infinitesimal. Other option is to check how it would change
with a finer frequency grid.